pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "myco"
dynamic = ["version"]
description = "Agent-First symbiotic cognitive substrate."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
  { name = "Yanjun" },
]
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
# ``pip install 'myco[mcp]'`` — pulls in the Model Context Protocol SDK
# so ``python -m myco.mcp`` (and any other MCP adapter) has what it needs.
# Floor of 1.2 covers the FastMCP.run() entry point used by the launcher;
# the low-level Server API has been stable across the entire 1.x line.
mcp = [
  "mcp>=1.2",
]
# ``pip install 'myco[adapters]'`` — pulls in optional deps for
# ingestion adapters: PDF, HTML, and URL. The built-in text-file,
# code-repo, CSV/JSON/JSONL, and tabular adapters use only the
# standard library and work without this extras target.
adapters = [
  "httpx>=0.27",
  "pypdf>=4.0",
  "beautifulsoup4>=4.12",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
# Three stable PATH-level entry points so `.mcp.json` and hook configs
# never have to reason about the user's venv or whether `python` vs
# `python3` is the right alias.
#
#   myco              — the manifest-driven CLI (12 verbs).
#   mcp-server-myco   — the MCP stdio/sse launcher; follows the
#                       community `mcp-server-<name>` convention so
#                       tools like uvx, pipx, and the MCP Registry
#                       can pick it up without configuration.
#   myco-install      — zero-copy-paste install into any of seven MCP
#                       hosts (Claude Code, Claude Desktop, Cursor,
#                       Windsurf, Zed, VS Code, OpenClaw). Writes the
#                       correct schema per host; idempotent.
myco = "myco.surface.cli:main"
mcp-server-myco = "myco.mcp:main"
myco-install = "myco.install:main"

# v0.5 (MAJOR 6) — dimension registration via entry-points.
#
# Myco's own built-in dimensions register here, the same way a third-
# party substrate plugin would. `default_registry()` discovers this
# table at runtime via `importlib.metadata.entry_points(group=
# "myco.dimensions")`. Adding a dimension means: write the class, add
# a row here, run `pip install -e .` (installed users get it
# automatically on next `pip install`). No edit to
# `src/myco/homeostasis/dimensions/__init__.py`'s `_BUILT_IN` tuple is
# strictly required after v0.5 — the tuple is the dev-checkout
# fallback, not the SSoT.
#
# Third-party dimension packages follow the same shape:
#
#   [project.entry-points."myco.dimensions"]
#   my_id = "mypkg.lint:MyDimensionClass"
[project.entry-points."myco.dimensions"]
M1 = "myco.homeostasis.dimensions.m1_canon_identity_fields:M1CanonIdentityFields"
M2 = "myco.homeostasis.dimensions.m2_entry_point_exists:M2EntryPointExists"
M3 = "myco.homeostasis.dimensions.m3_write_surface_declared:M3WriteSurfaceDeclared"
MF1 = "myco.homeostasis.dimensions.mf1_declared_subsystems_exist:MF1DeclaredSubsystemsExist"
SH1 = "myco.homeostasis.dimensions.sh1_package_version_ref:SH1PackageVersionRef"
MB1 = "myco.homeostasis.dimensions.mb1_raw_notes_backlog:MB1RawNotesBacklog"
MB2 = "myco.homeostasis.dimensions.mb2_no_integrated_yet:MB2NoIntegratedYet"
SE1 = "myco.homeostasis.dimensions.se1_dangling_refs:SE1DanglingRefs"
SE2 = "myco.homeostasis.dimensions.se2_orphan_integrated:SE2OrphanIntegrated"

[tool.hatch.version]
# Single source of truth for the package version — per L1 versioning.md.
path = "src/myco/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/myco"]

[tool.hatch.build.targets.sdist]
# Ephemeral / dev-only files that should never ship to PyPI even if
# they happen to sit at the repo root mid-release. The .gitignore
# mostly covers these already; this block is belt-and-suspenders so
# a stray `python -m build` while a release-notes scratch file is
# present does not leak it into the tarball.
include = [
  "/src",
  "/tests",
  "/docs",
  "/scripts",
  "/assets",
  "/skills",
  "/hooks",
  "/.claude-plugin",
  "/.claude",
  "/.mcp.json",
  "/_canon.yaml",
  "/MYCO.md",
  "/README.md",
  "/README_zh.md",
  "/README_ja.md",
  "/CHANGELOG.md",
  "/CONTRIBUTING.md",
  "/LICENSE",
  "/pyproject.toml",
  "/docs/INSTALL.md",
]
exclude = [
  "/.release_notes_*.md",
  "/upload*.log",
  "/dist",
  "/build",
  "/legacy_v0_3",
  "/.pytest_cache",
  "/.venv",
  "**/__pycache__",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
pythonpath = ["src"]
